Update main.py
I know, I know, it seems kind of pointless to update such a minuet amount of code, but I took a break from programming for a while, but I plan on getting back on it.
Basically, I fixed inefficient/ineffective and redundant lines that slowed the program; even if it was just a little bit. I also made the code recyclable -- as much as I could.